Move file system object to storage folder
